Remove redundant helpers from tests
