--- compose/foundation/foundation/src/jsNativeTest/kotlin/androidx/compose/foundation/internal/KeyEventUtils.jsNativeMain.kt
@@ -1,39 +1,0 @@
-/*
- * Copyright 2022 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package androidx.compose.foundation.internal
-
-import androidx.compose.ui.input.key.NativeKeyEvent
-import androidx.compose.ui.input.key.Key
-import androidx.compose.ui.input.key.KeyEvent
-import androidx.compose.ui.input.key.KeyEventType
-
-
-internal actual fun keyEvent(
-    key: Key,
-    keyEventType: KeyEventType,
-    modifiers: Int
-): KeyEvent {
-    return KeyEvent(
-        NativeKeyEvent(
-            key = key,
-            value = null,
-            modifiers = modifiers,
-            kind = keyEventType,
-            timestamp = 0L
-        )
-    )
-}
