+++ compose/ui/ui/src/desktopTest/kotlin/androidx/compose/ui/awt/ComposeWindowTest.kt
@@ -27,16 +27,13 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.input.pointer.PointerEventType
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.layout.layout
-import androidx.compose.ui.platform.WindowInfo
-import androidx.compose.ui.scene.BaseComposeScene
-import androidx.compose.ui.scene.ComposeScene
 import androidx.compose.ui.sendMouseEvent
 import androidx.compose.ui.sendMousePress
 import androidx.compose.ui.sendMouseRelease
-import androidx.compose.ui.window.WindowExceptionHandler
 import androidx.compose.ui.unit.Constraints
 import androidx.compose.ui.unit.IntSize
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.window.WindowExceptionHandler
 import androidx.compose.ui.window.density
 import androidx.compose.ui.window.runApplicationTest
 import com.google.common.truth.Truth.assertThat
@@ -293,9 +290,3 @@ class ComposeWindowTest {
 
     private class TestException : Exception()
 }
-
-private val ComposeScene.windowInfo: WindowInfo
-    get() {
-        this as BaseComposeScene
-        return composeSceneContext.platformContext.windowInfo
-    }
